fix potential overflow

+++ x/supply/keeper/keeper.go
@@ -82,7 +82,7 @@ func (k Keeper) CalculateCirculatingSupply(ctx sdk.Context, baseDenom string, ex
 	return k.bankKeeper.GetSupply(ctx, baseDenom).Amount.Sub(nonCirculating)
 }
 
-func (k Keeper) TopN(ctx sdk.Context, baseDenom string, n int) []*types.Account {
+func (k Keeper) TopN(ctx sdk.Context, baseDenom string, n uint64) []*types.Account {
 	accountMap := map[string]math.Int{}
 
 	modMap := map[string]bool{}
